+++ attn_gym/sparse/selected_attention/api.py
@@ -145,16 +145,15 @@ def selected_attention(
     attention_sink: Tensor,
     doc_ids: Tensor | None = None,
     sliding_window_size: int = 512,
-    share_kv: bool = True,
     backend: str = "triton",
     mode: str = "auto",
 ) -> Tensor:
     """
-    Performs selected attention
+    Performs selected attention.
         Each query attends to the previous sliding_window_size elements in the local_kv tensor
-        As well the positions in sparse_kv pointed to by kv_indices
-        Only one softmax is applied, and it covers the attention logits values generated by
-        the kv_indices as well as the sliding window
+        as well as the positions in sparse_kv pointed to by kv_indices.
+        Only one softmax is applied, covering both the sliding window and selected sparse positions.
+
     Args:
         query: query, shaped like (batch_size, num_heads, sequence_length, head_dim)
 
@@ -175,18 +174,16 @@ def selected_attention(
 
         doc_ids: Integer tensor in shape of (batch_size, sequence_length) or None.
             Looks something like [0, 0, 0, 1, 1, 2, 2, 2, 2], where tokens with the same id
-            can causally attend to each other
+            can causally attend to each other.
             If doc_ids[i, j] = doc_ids[i, j-y], then query[i, j] can causally attend to
-            local_kv[i, j-y]
-            Should be monotonically increasing on the sequence axis
+            local_kv[i, j-y].
+            Should be monotonically increasing on the sequence axis.
             If doc_ids is None, all tokens on the same sequence axis will be assumed to be in the
             same document.
             Only applies to the sliding window branch.
-            It's the caller's responsibility to make sure kv_indices don't cross document boundaries
+            It's the caller's responsibility to make sure kv_indices don't cross document boundaries.
 
         sliding_window_size: Integer, size of sliding window
-
-        share_kv: bool, true iff all query heads attend to the same KV head
 
         backend: one of eager, triton, or cute, controls which backend executes this code.
             Triton shared-KV backward uses nondeterministic atomic accumulation unless deterministic
@@ -237,7 +234,7 @@ def selected_attention(
                 share_kv,
             )
         case "cute":
-            from . import cute as cute_backend
+            from .impl import cute as cute_backend
 
             return cute_backend.selected_attention(
                 query,
